+++ src/frontend/apps/impress/src/features/docs/doc-management/hooks/index.ts
@@ -2,4 +2,5 @@ export * from './useCollaboration';
 export * from './useCopyDocLink';
 export * from './useDocUtils';
 export * from './useIsCollaborativeEditable';
+export * from './useMaskDocOption';
 export * from './useTrans';