+++ src/frontend/apps/impress/src/features/docs/docs-grid/components/DocsGridActions.tsx
@@ -9,6 +9,7 @@ import {
   useCreateFavoriteDoc,
   useDeleteFavoriteDoc,
   useDuplicateDoc,
+  useMaskDocOption,
 } from '@/docs/doc-management';
 
 interface DocsGridActionsProps {
@@ -31,6 +32,7 @@ export const DocsGridActions = ({
   const makeFavoriteDoc = useCreateFavoriteDoc({
     listInvalideQueries: [KEY_LIST_DOC],
   });
+  const maskDocOption = useMaskDocOption(doc);
 
   const options: DropdownMenuOption[] = [
     {
@@ -44,6 +46,7 @@ export const DocsGridActions = ({
         }
       },
       testId: `docs-grid-actions-${doc.is_favorite ? 'unpin' : 'pin'}-${doc.id}`,
+      showSeparator: true,
     },
     {
       label: t('Share'),
@@ -65,15 +68,21 @@ export const DocsGridActions = ({
           canSave: false,
         });
       },
-    },
-    {
-      label: t('Remove'),
-      icon: 'delete',
-      callback: () => deleteModal.open(),
-      disabled: !doc.abilities.destroy,
-      testId: `docs-grid-actions-remove-${doc.id}`,
+      showSeparator: true,
     },
   ];
+
+  const leaveDocOption: DropdownMenuOption = doc.abilities.destroy
+    ? {
+        label: t('Delete document'),
+        icon: 'delete',
+        callback: () => deleteModal.open(),
+        disabled: !doc.abilities.destroy,
+        testId: `docs-grid-actions-remove-${doc.id}`,
+      }
+    : maskDocOption;
+
+  options.push(leaveDocOption);
 
   return (
     <>